Added basic location hash implementation.
The format's new getSlug method returns a slug version of the format's name. This will be used to update and recognize formats inside the url or hash provieded. Format options that hold non-default values will be included in the url or hash. Option updates cause a location (url or hash) update.