Added an animated logo to the side view.
Moved tick timer from DeckView to Application. Implemented the Page Visibility api that clears the tick timer when the page or tab is not active. The SideView only handle and forwards ticks when it is visible. Introduced a new LogoView that draws and animates the logo using a canvas element. It has been added to the SideView. Animations are randomly triggered by the tick timer.